+++ BranchAndPrune.hpp
@@ -10,14 +10,14 @@
 class BranchAndPrune : public Solveur {
 
 public:
-	BranchAndPrune(int x, std::vector<Constraint*> contraintes);
+	BranchAndPrune(unsigned short int x, std::vector<Constraint*> contraintes);
 	int solve();
-    std::vector<Noeud> branch(Noeud noeud);
+    	std::vector<Noeud> branch(Noeud noeud);
 
 private:
 
     int solveRec(Noeud noeud);
-    void prune(int val, Noeud *n);
+    void prune(unsigned short int val, Noeud *n);
 
     Problem problem;
