+++ BranchAndPruneNonRec.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-BranchAndPruneNonRec::BranchAndPruneNonRec(int nombre_dames, std::vector<Constraint*> contraintes):problem(nombre_dames,contraintes){
+BranchAndPruneNonRec::BranchAndPruneNonRec(unsigned short int nombre_dames, std::vector<Constraint*> contraintes):problem(nombre_dames,contraintes){
 	noeuds.push_front(problem.initialNode());
 }
 
@@ -60,8 +60,8 @@ void BranchAndPruneNonRec::branch(Noeud noeud){
 	while(!trouve){
 		//si le domaine a la position i a une taille superieure a 1 alors on peut generer les sous noeuds
 		if(noeud.getDomains().at(i).size() > 1){
-			set<int> domain = noeud.getDomains().at(i);
-			for (std::set<int>::iterator it = domain.begin(); it != domain.end(); it++){
+			set<unsigned short int> domain = noeud.getDomains().at(i);
+			for (std::set<unsigned short int>::iterator it = domain.begin(); it != domain.end(); it++){
 				Noeud n;
 				n = noeud.copie();
 				//on supprime la valeur choisie dans tous les ensembles
@@ -75,7 +75,7 @@ void BranchAndPruneNonRec::branch(Noeud noeud){
 	}
 }
 
-void BranchAndPruneNonRec::supprimer(int val,Noeud *n){
+void BranchAndPruneNonRec::supprimer(unsigned short int val,Noeud *n){
 	for (int i = 0; i < n->getDomains().size(); i++){
 		n->supprimer(i, val);
 	}
