+++ Noeud.cpp
@@ -7,8 +7,8 @@ Noeud::Noeud(){
 }
 
 
-Noeud::Noeud(int val){
-	std::set<int> domain;
+Noeud::Noeud(unsigned short int val){
+	std::set<unsigned short int> domain;
 	for(int i = 1; i <= val; i++){
 		domain.insert(i);
 	}
@@ -18,7 +18,7 @@ Noeud::Noeud(int val){
 }
 
 void Noeud::toString(){
-	std::set<int>::iterator it;
+	std::set<unsigned short int>::iterator it;
 	for(int j = 0; j < domains.size(); j++){
 		cout<<"Domain "<<j+1<<": ";
 		for(it = domains.at(j).begin(); it != domains.at(j).end(); ++it){
@@ -30,7 +30,7 @@ void Noeud::toString(){
 }
 
 void Noeud::toString2(){
-	std::set<int>::iterator it;
+	std::set<unsigned short int>::iterator it;
 	for(int j = 0; j < domains.size(); j++){
 		for(it = domains.at(j).begin(); it != domains.at(j).end(); ++it){
 			cout<<*it<<", ";
@@ -39,7 +39,7 @@ void Noeud::toString2(){
 	cout<<endl;
 }
 
-std::vector<std::set<int> > Noeud::getDomains(){
+std::vector<std::set<unsigned short int> > Noeud::getDomains(){
 	return domains;
 }
 
@@ -49,7 +49,7 @@ int Noeud::taille(){
 }
 
 
-void Noeud::setDomains(std::vector<std::set<int> > new_domains){
+void Noeud::setDomains(std::vector<std::set<unsigned short int> > new_domains){
 	domains = new_domains;
 }
 
@@ -59,19 +59,19 @@ Noeud Noeud::copie(){
 	return n;
 }
 
-void Noeud::clear_and_add(int pos, int val){
+void Noeud::clear_and_add(int pos, unsigned short int val){
 	domains.at(pos).clear();
 	domains.at(pos).insert(val);
 }
 
-void Noeud::supprimer(int pos, int val){
+void Noeud::supprimer(int pos, unsigned short int val){
 	domains.at(pos).erase(val);
 }
 
 
 
 void Noeud::move(int i){
-	std::set<int> temp = domains.at((i+ 1)%domains.size() );
+	std::set<unsigned short int> temp = domains.at((i+ 1)%domains.size() );
 	domains.at((i+ 1)%domains.size() ) = domains.at(i%domains.size());
 	domains.at(i%domains.size()) = temp;
 }
