+++ Nqueen_colonne.cpp
@@ -7,12 +7,12 @@ Nqueen_colonne::Nqueen_colonne(){
 }
 
 Proof Nqueen_colonne::test(Noeud noeud){
-	std::vector<std::set<int> > domains = noeud.getDomains();
+	std::vector<std::set<unsigned short int> > domains = noeud.getDomains();
 	//pour savoir quelles colonnes sont deja utilises dans un noeud
-	set<int> colonnes_utilises;
+	set<unsigned short int> colonnes_utilises;
 	Proof resultat = succes;
 	for(int i = 0; i < domains.size(); i++){
-		std::set<int>::iterator it;
+		std::set<unsigned short int>::iterator it;
 
 		if(domains.at(i).size() == 1){
 			it = colonnes_utilises.find(*domains.at(i).begin());
